+++ tests/Support/Model/DefaultValueChild.php
@@ -4,15 +4,13 @@ declare(strict_types=1);
 
 namespace UIAwesome\Model\Tests\Support\Model;
 
-use UIAwesome\Model\AbstractModel;
-
 /**
- * Stub container model for nested trim for tests.
+ * Stub child model reusing a parent property name for tests.
  *
  * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-final class TrimContainer extends AbstractModel
+final class DefaultValueChild extends DefaultValueParent
 {
-    public function __construct(public readonly TrimAddress $address) {}
+    public string|null $status = '';
 }