+++ tests/Support/Model/Dynamic.php
@@ -7,7 +7,7 @@ namespace UIAwesome\Model\Tests\Support\Model;
 use UIAwesome\Model\AbstractModel;
 
 /**
- * Stub dynamic model used by test fixtures.
+ * Stub dynamic model used for tests.
  *
  * @copyright Copyright (C) 2024 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.