+++ tests/Support/Model/MapFromDuplicate.php
@@ -8,7 +8,7 @@ use UIAwesome\Model\AbstractModel;
 use UIAwesome\Model\Attribute\MapFrom;
 
 /**
- * Stub model with duplicate MapFrom keys used by tests.
+ * Stub model with duplicate MapFrom keys used for tests.
  *
  * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.