+++ tests/Support/Model/NoSnakeCasePayload.php
@@ -8,7 +8,7 @@ use UIAwesome\Model\AbstractModel;
 use UIAwesome\Model\Attribute\NoSnakeCase;
 
 /**
- * Stub model containing NoSnakeCase-marked properties.
+ * Stub model containing NoSnakeCase-marked properties used for tests.
  *
  * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.