+++ tests/Support/Model/NonNamespaced.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 use UIAwesome\Model\AbstractModel;
 
 /**
- * Stub non-namespaced model used by test fixtures.
+ * Stub non-namespaced model used for tests.
  *
  * @copyright Copyright (C) 2024 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.