+++ tests/Support/Model/Profile.php
@@ -8,7 +8,7 @@ use UIAwesome\Model\AbstractModel;
 use UIAwesome\Model\Attribute\DoNotCollect;
 
 /**
- * Stub profile model with nested address relation used by test fixtures.
+ * Stub profile model with nested address relation used for tests.
  *
  * @copyright Copyright (C) 2024 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.